Rename : 프로젝트 폴더 FSD구조화
- 프로젝트 기능별로 응집도를 분리하기 위해 FSD 구조로 변경하였습니다.

ref : #1
