+++ src/app/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Header from './components/layout/Header';
-import Home from './pages/Home';
+import Header from '../widgets/layout/Header';
+import Home from '../pages/Home';
 
 function App() {
   return (
