+++ src/features/home/api/index.ts
@@ -1,0 +1,1 @@
+/** index.ts */
