+++ src/features/home/ui/StepSection.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import theme from '../../styles/theme';
+import theme from '../../../shared/styles/theme';
 
 type StepInfo = {
   step: number; // 단계
