+++ src/pages/Home.tsx
@@ -1,11 +1,11 @@
 import { useRef } from 'react';
 import styled from '@emotion/styled';
 import { Fade } from 'react-awesome-reveal';
-import { FirstSection } from '../components/home/FirstSection';
-import { SecondSection } from '../components/home/SecondSection';
-import { ThirdSection } from '../components/home/ThirdSection';
-import { FourthSection } from '../components/home/FourthSection';
-import { FifthSection } from '../components/home/FifthSection';
+import { FirstSection } from '../features/home/ui/FirstSection';
+import { SecondSection } from '../features/home/ui/SecondSection';
+import { ThirdSection } from '../features/home/ui/ThirdSection';
+import { FourthSection } from '../features/home/ui/FourthSection';
+import { FifthSection } from '../features/home/ui/FifthSection';
 
 const Home = () => {
   const scrollDownRef = useRef<HTMLDivElement>(null); // 스크롤 섹션 참조
