+++ src/shared/assets/font/Pretendard-Medium.otf
