+++ src/shared/assets/font/Pretendard-Medium.ttf
