+++ src/shared/data/index.ts
@@ -1,0 +1,1 @@
+/** 공통데이터  */
