+++ src/shared/hooks/useDevice.ts
