+++ src/shared/store/index.ts
