--- src/utils/index.ts
@@ -1,1 +1,0 @@
-// index.ts
